+++ hotfixes-service-springboot-starter/src/main/java/com/github/jtrinity/dbc/domain/WmoAreaTable.java
@@ -69,7 +69,7 @@ public class WmoAreaTable implements DbcEntity {
     private Byte soundProviderPrefUnderwater;
 
     @Column(name = "Flags")
-    private Byte flags;
+    private Short flags;
 
     @Id
     @ColumnDefault("'0'")
